requirements.txt
# Telegram
aiogram>=3.3.0

# Конфигурация
python-dotenv>=1.0.0

# HTTP
aiohttp>=3.9.0

# PDF
reportlab>=4.0.0

# AI
openai>=1.3.0

# Тесты
pytest>=7.4.0
pytest-asyncio>=0.21.0
